+++ orury-domain/src/main/java/org/orury/domain/user/domain/UserServiceImpl.java
@@ -1,4 +1,4 @@
-package org.orury.client.user.service;
+package org.orury.domain.user.domain;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -14,9 +14,8 @@ import org.orury.domain.global.domain.ImageUtils;
 import org.orury.domain.gym.domain.GymStore;
 import org.orury.domain.review.db.repository.ReviewReactionRepository;
 import org.orury.domain.review.db.repository.ReviewRepository;
-import org.orury.domain.user.db.model.User;
-import org.orury.domain.user.db.repository.UserRepository;
-import org.orury.domain.user.dto.UserDto;
+import org.orury.domain.user.domain.dto.UserDto;
+import org.orury.domain.user.domain.entity.User;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -24,8 +23,9 @@ import org.springframework.web.multipart.MultipartFile;
 @Slf4j
 @RequiredArgsConstructor
 @Service
-public class UserService {
-    private final UserRepository userRepository;
+public class UserServiceImpl implements UserService {
+    private final UserReader userReader;
+    private final UserStore userStore;
     private final ImageUtils imageUtils;
     private final PostRepository postRepository;
     private final PostLikeRepository postLikeRepository;
@@ -35,13 +35,14 @@ public class UserService {
     private final ReviewReactionRepository reviewReactionRepository;
     private final GymStore gymStore;
 
+    @Override
     public UserDto getUserDtoById(Long id) {
-        User user = userRepository.findById(id)
-                .orElseThrow(() -> new BusinessException(UserErrorCode.NOT_FOUND));
+        User user = userReader.findUserById(id);
         String profileUrl = imageUtils.getUserImageUrl(user.getProfileImage());
         return UserDto.from(user, profileUrl);
     }
 
+    @Override
     @Transactional
     public void updateProfileImage(UserDto userDto, MultipartFile image) throws BusinessException {
         //기존에 저장된 요소 삭제
@@ -50,11 +51,13 @@ public class UserService {
         imageUploadAndSave(userDto, image);
     }
 
+    @Override
     @Transactional
     public void updateUserInfo(UserDto userDto) {
-        userRepository.save(userDto.toEntity(ImageUrlConverter.splitUrlToImage(userDto.profileImage())));
+        userStore.save(userDto.toEntity(ImageUrlConverter.splitUrlToImage(userDto.profileImage())));
     }
 
+    @Override
     @Transactional
     public void deleteUser(UserDto userDto) {
         deleteReviewReactionsByUserId(userDto.id());
@@ -66,7 +69,7 @@ public class UserService {
         imageUtils.oldS3ImagesDelete(S3Folder.USER.getName(), userDto.profileImage());
         var deletingUser = userDto.toEntity().delete(imageUtils.getUserDefaultImage());
 
-        userRepository.save(deletingUser);
+        userStore.save(deletingUser);
     }
 
     private void deleteReviewReactionsByUserId(Long userId) {
@@ -115,11 +118,11 @@ public class UserService {
     private void imageUploadAndSave(UserDto userDto, MultipartFile file) {
         //MultipartFile이 오지 않은 경우 -> 유저의 프로필 이미지를 기본 이미지로 변경
         if (file == null || file.isEmpty()) {
-            userRepository.save(userDto.toEntity(imageUtils.getUserDefaultImage()));
+            userStore.save(userDto.toEntity(imageUtils.getUserDefaultImage()));
         } else {
             //MultipartFile이 오는 경우 -> 유저의 프로필 이미지를 업로드한 이미지로 변경
             String image = imageUtils.upload(S3Folder.USER.getName(), file);
-            userRepository.save(userDto.toEntity(image));
+            userStore.save(userDto.toEntity(image));
         }
     }
 }
